Feature/cor 1498 vaccinations toggle button removal cherrypick (#4728)
* fix: merged conflicts

* fix: removed left vr

* feat: merged two files together

* Revert "fix: merged conflicts"

This reverts commit 8a14ed0e332d8c01af187685bc96d584be11f492, reversing
changes made to 5a152ffb35e39c0738f83cd1bd9f75b99bba2920.

* fix: merged conflicts

* fix: removed left vr

* feat: merged two files together

* fix: conflict

---------
