+++ packages/app/src/domain/vaccine/vaccine-coverage-choropleth.tsx
@@ -1,12 +1,11 @@
-import { colors, GmCollectionVaccineCoveragePerAgeGroup, VrCollectionVaccineCoveragePerAgeGroup } from '@corona-dashboard/common';
+import { colors, GmCollectionVaccineCoveragePerAgeGroup } from '@corona-dashboard/common';
 import { SiteText } from '~/locale';
 import { matchingAgeGroups, VaccineCoverageData, DataPerAgeGroup, BirthyearRangeKeysOfAgeGroups, PercentageKeysOfAgeGroups, PercentageLabelKeysOfAgeGroups } from './common';
 import css from '@styled-system/css';
 import { useState } from 'react';
 import { space } from '~/style/theme';
 import { Box } from '~/components/base';
-import { RegionControlOption } from '~/components/chart-region-controls';
-import { DynamicChoropleth } from '~/components/choropleth';
+import { DataOptions, DynamicChoropleth } from '~/components/choropleth';
 import { ChoroplethTile } from '~/components/choropleth-tile';
 import { thresholds } from '~/components/choropleth/logic';
 import { TooltipContent, TooltipSubject } from '~/components/choropleth/tooltips';
@@ -15,25 +14,23 @@ import { Markdown } from '~/components/markdown';
 import { BoldText } from '~/components/typography';
 import { useIntl } from '~/intl';
 import { replaceVariablesInText } from '~/utils/replace-variables-in-text';
-import { useReverseRouter } from '~/utils/use-reverse-router';
 import { AgeGroup, AgeGroupSelect } from './components/age-group-select';
 import { CoverageKindProperty, VaccinationCoverageKindSelect } from './components/vaccination-coverage-kind-select';
 import { parseVaccinatedPercentageLabel } from './logic/parse-vaccinated-percentage-label';
 
 interface VaccineCoverageChoroplethProps {
-  data: {
-    gm: GmCollectionVaccineCoveragePerAgeGroup[];
-    vr: VrCollectionVaccineCoveragePerAgeGroup[];
+  data: GmCollectionVaccineCoveragePerAgeGroup[];
+  dataOptions: DataOptions;
+  text: {
+    title: string;
+    description: string;
   };
 }
 
-export const VaccineCoverageChoropleth = ({ data }: VaccineCoverageChoroplethProps) => {
+export const VaccineCoverageChoropleth = ({ data, dataOptions, text }: VaccineCoverageChoroplethProps) => {
   const { commonTexts } = useIntl();
-  const [selectedMap, setSelectedMap] = useState<RegionControlOption>('gm');
   const [selectedAgeGroup, setSelectedAgeGroup] = useState<AgeGroup>('18');
   const [selectedCoverageKind, setSelectedCoverageKind] = useState<CoverageKindProperty>('primary_series');
-  const reverseRouter = useReverseRouter();
-
   /**
    * When changing between coverage kinds where the selected age group isn't available,
    * the other coverage kind set the non-matching age group to a default one.
@@ -46,19 +43,14 @@ export const VaccineCoverageChoropleth = ({ data }: VaccineCoverageChoroplethPro
     setSelectedCoverageKind(coverageKind);
   };
 
-  const variables = {
-    regio: commonTexts.choropleth.choropleth_vaccination_coverage.shared[selectedMap],
-  };
-
-  const choroplethDataVr: VrCollectionVaccineCoveragePerAgeGroup[] = data.vr.filter((choroplethDataSingleVR) => choroplethDataSingleVR.vaccination_type === selectedCoverageKind);
-  const choroplethDataGm: GmCollectionVaccineCoveragePerAgeGroup[] = data.gm.filter((choroplethDataSingleGM) => choroplethDataSingleGM.vaccination_type === selectedCoverageKind);
+  const choroplethDataGm: GmCollectionVaccineCoveragePerAgeGroup[] = data.filter((choroplethDataSingleGM) => choroplethDataSingleGM.vaccination_type === selectedCoverageKind);
 
   return (
     <ChoroplethTile
-      title={replaceVariablesInText(commonTexts.choropleth.choropleth_vaccination_coverage.nl.title, variables)}
+      title={text.title}
       description={
         <>
-          <Markdown content={replaceVariablesInText(commonTexts.choropleth.choropleth_vaccination_coverage.nl.description, variables)} />
+          <Markdown content={text.description} />
           <Box display="flex" flexDirection="row" justifyContent="flex-start" spacingHorizontal={2} as={'fieldset'}>
             <BoldText
               as="legend"
@@ -82,51 +74,26 @@ export const VaccineCoverageChoropleth = ({ data }: VaccineCoverageChoroplethPro
         </>
       }
       legend={{
-        thresholds: selectedMap === 'gm' ? thresholds.gm.primary_series_percentage : thresholds.vr.primary_series_percentage,
+        thresholds: thresholds.gm.primary_series_percentage,
         title: commonTexts.choropleth.choropleth_vaccination_coverage.shared.legend_title,
       }}
       metadata={{
         source: commonTexts.choropleth.vaccination_coverage.shared.bronnen.rivm,
-        date:
-          selectedMap === 'gm'
-            ? data.gm.find((item: GmCollectionVaccineCoveragePerAgeGroup | VrCollectionVaccineCoveragePerAgeGroup) => item.vaccination_type === selectedCoverageKind)?.date_unix
-            : data.vr.find((item: GmCollectionVaccineCoveragePerAgeGroup | VrCollectionVaccineCoveragePerAgeGroup) => item.vaccination_type === selectedCoverageKind)?.date_unix,
+        date: data.find((item: GmCollectionVaccineCoveragePerAgeGroup) => item.vaccination_type === selectedCoverageKind)?.date_unix,
       }}
-      chartRegion={selectedMap}
-      onChartRegionChange={setSelectedMap}
+      hasPadding
     >
-      {selectedMap === 'gm' && (
-        <DynamicChoropleth
-          map={'gm'}
-          accessibility={{ key: 'vaccine_coverage_nl_choropleth' }}
-          data={choroplethDataGm}
-          dataConfig={{
-            metricName: 'vaccine_coverage_per_age_group',
-            metricProperty: `vaccinated_percentage_${selectedAgeGroup}_plus`,
-          }}
-          dataOptions={{
-            isPercentage: true,
-            getLink: (gmcode) => reverseRouter.gm.vaccinaties(gmcode),
-          }}
-          formatTooltip={(context) => <ChoroplethTooltip data={context} ageGroups={matchingAgeGroups[selectedCoverageKind]} selectedCoverageKind={selectedCoverageKind} />}
-        />
-      )}
-
-      {selectedMap === 'vr' && (
-        <DynamicChoropleth
-          map={'vr'}
-          accessibility={{ key: 'vaccine_coverage_nl_choropleth' }}
-          data={choroplethDataVr}
-          dataConfig={{
-            metricName: 'vaccine_coverage_per_age_group',
-            metricProperty: `vaccinated_percentage_${selectedAgeGroup}_plus`,
-          }}
-          dataOptions={{
-            isPercentage: true,
-          }}
-          formatTooltip={(context) => <ChoroplethTooltip data={context} ageGroups={matchingAgeGroups[selectedCoverageKind]} selectedCoverageKind={selectedCoverageKind} />}
-        />
-      )}
+      <DynamicChoropleth
+        map={'gm'}
+        accessibility={{ key: 'vaccine_coverage_nl_choropleth' }}
+        data={choroplethDataGm}
+        dataConfig={{
+          metricName: 'vaccine_coverage_per_age_group',
+          metricProperty: `vaccinated_percentage_${selectedAgeGroup}_plus`,
+        }}
+        dataOptions={dataOptions}
+        formatTooltip={(context) => <ChoroplethTooltip data={context} ageGroups={matchingAgeGroups[selectedCoverageKind]} selectedCoverageKind={selectedCoverageKind} />}
+      />
     </ChoroplethTile>
   );
 };
